Refactor Laravel php unit tests
