+++ tests/Integrations/Laravel/V4/PathParamsTest.php
@@ -2,53 +2,15 @@
 
 namespace DDTrace\Tests\Integrations\Laravel\V4;
 
-use DDTrace\Tests\Common\AppsecTestCase;
-use DDTrace\Tests\Frameworks\Util\Request\GetSpec;
-use datadog\appsec\AppsecStatus;
+use DDTrace\Tests\Integrations\Laravel\PathParamsTestSuite;
 
 /**
  * @group appsec
  */
-class PathParamsTest extends AppsecTestCase
+class PathParamsTest extends PathParamsTestSuite
 {
     protected static function getAppIndexScript()
     {
         return __DIR__ . '/../../../Frameworks/Laravel/Version_4_2/public/index.php';
-    }
-
-    public function testDynamicRouteWithAllParametersGiven()
-    {
-        $param01 = 'first_param';
-        $param02 = 'second_param';
-        $this->call(
-            GetSpec::create('Call to dynamic route', "/dynamic_route/$param01/static/$param02")
-        );
-        $events = AppsecStatus::getInstance()->getEvents();
-        $this->assertEquals(1, count($events));
-        $this->assertEquals($param01, $events[0]["server.request.path_params"]['param01']);
-        $this->assertEquals($param02, $events[0]["server.request.path_params"]['param02']);
-        $this->assertEquals('push_address', $events[0]['eventName']);
-    }
-
-    public function testDynamicRouteWithOptionalParametersNotGiven()
-    {
-        $param01 = 'first_param';
-        $this->call(
-            GetSpec::create('Call to dynamic route', "/dynamic_route/$param01/static")
-        );
-        $events = AppsecStatus::getInstance()->getEvents();
-        $this->assertEquals(1, count($events));
-        $this->assertCount(1, $events[0]["server.request.path_params"]);
-        $this->assertEquals($param01, $events[0]["server.request.path_params"]['param01']);
-        $this->assertEquals('push_address', $events[0]['eventName']);
-    }
-
-    public function testStaticRouteDoesNotGenerateEvent()
-    {
-        $this->call(
-            GetSpec::create('Call to static route', "/simple")
-        );
-        $events = AppsecStatus::getInstance()->getEvents();
-        $this->assertEquals(0, count($events));
     }
 }
